Break up duplicated dropdown contents
